pyproject.toml
[tool.poetry]
name="spotify-playlist-generator"
version="0.1.0"
description="Spotify playlist generator"
authors = ["NTTLuke"]

[tool.poetry.dependencies]
python = ">=3.10.0,<=3.12.1"
crewai = { extras = ["tools"], version = "*" }
unstructured = '==0.10.25'
pyowm = '3.3.0'
python-dotenv = "1.0.0"
setuptools = "^69.1.0"
fastapi = "^0.109.2"
uvicorn = "^0.28.0"


[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"